Update France.json
Relations 349015 and 349042 are relations in Spain but overlap with the disputed island between France and Spain which is controlled by each country for half of a year. 

Relation 45155 is controlled by Italy but has a small overlap with France on the disputed part of Alpes. 

Other relations are France's overseas territories.