Update China.json
Local user insist that Pingtan is a separate unit, but it has some special status which is not clear. It should not be a part of existing admin_level=5 relation, but local users claim this to be a same level of hierarchy special unit. Adding it to the list as expected state